refactory: add handle multiple promises function
